requirements.txt
django
djangorestframework
django_extensions
